refactor: collapse test files to make it less verbose
