--- kv/providers/gcp/external_account_internal_test.go
@@ -1,153 +1,0 @@
-package gcp
-
-import (
-	"os"
-	"path/filepath"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-const validWIF = `{"type":"external_account",` +
-	`"token_url":"https://sts.googleapis.com/v1/token",` +
-	`"credential_source":{"file":"/var/run/token"}}`
-
-func TestValidateExternalAccount(t *testing.T) {
-	t.Parallel()
-
-	const googleImpersonationURL = `"https://iamcredentials.googleapis.com/v1/projects/-/` +
-		`serviceAccounts/x@y.iam.gserviceaccount.com:generateAccessToken"`
-
-	tests := []struct {
-		name        string
-		json        string
-		wantErr     bool
-		errContains string
-	}{
-		{
-			name: "valid file source",
-			json: validWIF,
-		},
-		{
-			name: "valid with Google impersonation url",
-			json: `{"type":"external_account",` +
-				`"token_url":"https://sts.googleapis.com/v1/token",` +
-				`"service_account_impersonation_url":` + googleImpersonationURL + `,` +
-				`"credential_source":{"file":"/t"}}`,
-		},
-		{
-			name: "wrong type",
-			json: `{"type":"service_account",` +
-				`"token_url":"https://sts.googleapis.com/v1/token","credential_source":{"file":"/t"}}`,
-			wantErr:     true,
-			errContains: "type",
-		},
-		{
-			name: "non-Google token_url",
-			json: `{"type":"external_account",` +
-				`"token_url":"https://evil.example.com/token","credential_source":{"file":"/t"}}`,
-			wantErr:     true,
-			errContains: "token_url",
-		},
-		{
-			name: "non-Google impersonation url",
-			json: `{"type":"external_account",` +
-				`"token_url":"https://sts.googleapis.com/v1/token",` +
-				`"service_account_impersonation_url":"https://evil.example.com/x:generateAccessToken",` +
-				`"credential_source":{"file":"/t"}}`,
-			wantErr:     true,
-			errContains: "impersonation",
-		},
-		{
-			name: "executable credential source rejected",
-			json: `{"type":"external_account",` +
-				`"token_url":"https://sts.googleapis.com/v1/token",` +
-				`"credential_source":{"executable":{"command":"/bin/evil"}}}`,
-			wantErr:     true,
-			errContains: "executable",
-		},
-		{
-			name:        "malformed json",
-			json:        `{not json`,
-			wantErr:     true,
-			errContains: "invalid external_account",
-		},
-		{
-			name: "valid aws imds source",
-			json: `{"type":"external_account",` +
-				`"token_url":"https://sts.googleapis.com/v1/token",` +
-				`"credential_source":{"environment_id":"aws1",` +
-				`"region_url":"http://169.254.169.254/latest/meta-data/placement/availability-zone",` +
-				`"url":"http://169.254.169.254/latest/meta-data/iam/security-credentials",` +
-				`"imdsv2_session_token_url":"http://169.254.169.254/latest/api/token"}}`,
-		},
-		{
-			name: "valid aws imds source over ipv6",
-			json: `{"type":"external_account",` +
-				`"token_url":"https://sts.googleapis.com/v1/token",` +
-				`"credential_source":{"environment_id":"aws1",` +
-				`"url":"http://[fd00:ec2::254]/latest/meta-data/iam/security-credentials"}}`,
-		},
-		{
-			name: "tampered aws credential source url",
-			json: `{"type":"external_account",` +
-				`"token_url":"https://sts.googleapis.com/v1/token",` +
-				`"credential_source":{"environment_id":"aws1",` +
-				`"url":"http://evil.example.com/latest/meta-data/iam/security-credentials"}}`,
-			wantErr:     true,
-			errContains: "IMDS",
-		},
-		{
-			name: "non-aws url source is not imds-restricted",
-			json: `{"type":"external_account",` +
-				`"token_url":"https://sts.googleapis.com/v1/token",` +
-				`"credential_source":{"url":"https://my-idp.example.com/token"}}`,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			err := validateExternalAccount(&Config{
-				CredentialsType: "external_account",
-				CredentialsJSON: tt.json,
-			})
-
-			if tt.wantErr {
-				require.Error(t, err)
-				require.ErrorContains(t, err, tt.errContains)
-
-				return
-			}
-
-			require.NoError(t, err)
-		})
-	}
-}
-
-func TestValidateExternalAccount_FileSource(t *testing.T) {
-	t.Parallel()
-
-	t.Run("reads and validates from credentials_file", func(t *testing.T) {
-		t.Parallel()
-
-		path := filepath.Join(t.TempDir(), "wif.json")
-		require.NoError(t, os.WriteFile(path, []byte(validWIF), 0o600))
-
-		require.NoError(t, validateExternalAccount(&Config{
-			CredentialsType: "external_account",
-			CredentialsFile: path,
-		}))
-	})
-
-	t.Run("unreadable credentials_file is an error", func(t *testing.T) {
-		t.Parallel()
-
-		err := validateExternalAccount(&Config{
-			CredentialsType: "external_account",
-			CredentialsFile: filepath.Join(t.TempDir(), "missing.json"),
-		})
-		require.Error(t, err)
-	})
-}
